frontend/requirements.txt
streamlit==1.45.0
httpx==0.28.1
python-dotenv==1.1.0
loguru==0.7.3
